[#3] Change field 'id' from long type to primitive int

+++ src/main/java/com/thoughtworks/gradle/gia/domain/TodoItem.java
@@ -1,15 +1,15 @@
 package com.thoughtworks.gradle.gia.domain;
 
 public class TodoItem implements Comparable<TodoItem> {
-    private long id;
+    private int id;
     private String name;
     private boolean completed;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public TodoItem id(long id) {
+    public TodoItem id(int id) {
         this.id = id;
         return this;
     }
